requirements.txt
asgiref==3.2.10
Django==3.1
django-crispy-forms==1.9.2
django-filter==2.3.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
